poll/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
astroid==1.5.3
cffi==1.10.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
cryptography==2.0.3
dj-database-url==0.4.2
Django==1.11.2
django-cors-headers==2.1.0
djangorestframework==3.6.3
gufw==17.10.0
gunicorn==19.7.1
idna==2.5
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector==2.1.4
mysqlclient==1.3.10
npyscreen==4.10.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pacman-mirrors==4.2.0
Pillow==4.2.1
ply==3.10
pycparser==2.18
pygobject==3.24.1
pylint==1.7.2
pyparsing==2.2.0
pysrt==1.1.1
pytz==2017.2
pyxdg==0.25
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
team==1.0
urllib3==1.22
virtualenv==15.1.0
wrapt==1.10.10
